[common/meta/api] refactor: move test to crate "meta/api" so that meta-embedded and meta-srv reuse the test suite
