+++ common/meta/api/Cargo.toml
@@ -9,10 +9,15 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+common-datavalues = {path = "../../datavalues" }
 common-exception = {path = "../../exception" }
-common-planners = {path = "../../planners"}
 common-meta-types = {path= "../types" }
+common-planners = {path = "../../planners"}
+common-tracing = { path = "../../tracing" }
 
 
-futures = "0.3"
+
+anyhow = "1.0.44"
 async-trait = "0.1"
+futures = "0.3"
+maplit = "1.0.2"
